requirements.txt
flask
flask-wtf
# flask-uploads
flask-reuploaded
pillow
pexpect
flask-sockets
